requirements.txt
django
djangorestframework
flake8
docutils
psycopg2
Pillow